+++ src/host/configuration.h
@@ -114,14 +114,14 @@ namespace host
     struct Command
     {
       StartType type = StartType::Start;
-      std::string network_certificate_file = "networkcert.pem";
+      std::string service_certificate_file = "service_cert.pem";
 
       struct Start
       {
         std::vector<ParsedMemberInfo> members = {};
         std::vector<std::string> constitution_files = {};
         ccf::ServiceConfiguration service_configuration;
-        size_t initial_network_certificate_validity_days = 1;
+        size_t initial_service_certificate_validity_days = 1;
 
         bool operator==(const Start&) const = default;
       };
@@ -138,7 +138,7 @@ namespace host
 
       struct Recover
       {
-        size_t initial_network_certificate_validity_days = 1;
+        size_t initial_service_certificate_validity_days = 1;
         bool operator==(const Recover&) const = default;
       };
       Recover recover = {};
@@ -182,7 +182,7 @@ namespace host
   DECLARE_JSON_OPTIONAL_FIELDS(
     CCHostConfig::Command::Start,
     service_configuration,
-    initial_network_certificate_validity_days);
+    initial_service_certificate_validity_days);
 
   DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(CCHostConfig::Command::Join);
   DECLARE_JSON_REQUIRED_FIELDS(CCHostConfig::Command::Join, target_rpc_address);
@@ -191,12 +191,12 @@ namespace host
   DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(CCHostConfig::Command::Recover);
   DECLARE_JSON_REQUIRED_FIELDS(CCHostConfig::Command::Recover);
   DECLARE_JSON_OPTIONAL_FIELDS(
-    CCHostConfig::Command::Recover, initial_network_certificate_validity_days);
+    CCHostConfig::Command::Recover, initial_service_certificate_validity_days);
 
   DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(CCHostConfig::Command);
   DECLARE_JSON_REQUIRED_FIELDS(CCHostConfig::Command, type);
   DECLARE_JSON_OPTIONAL_FIELDS(
-    CCHostConfig::Command, network_certificate_file, start, join, recover);
+    CCHostConfig::Command, service_certificate_file, start, join, recover);
 
   DECLARE_JSON_TYPE_WITH_BASE_AND_OPTIONAL_FIELDS(CCHostConfig, CCFConfig);
   DECLARE_JSON_REQUIRED_FIELDS(CCHostConfig, enclave, command);