+++ src/node/node_to_node.h
@@ -97,7 +97,7 @@ namespace ccf
 
     virtual void initialize(
       const NodeId& self_id,
-      const crypto::Pem& network_cert,
+      const crypto::Pem& service_cert,
       crypto::KeyPairPtr node_kp,
       const std::optional<crypto::Pem>& node_cert = std::nullopt) = 0;